Refactored the fish classes

+++ src/main/scala/biology/Ontogeny.scala
@@ -4,6 +4,6 @@ import locals.OntogenyState.OntogenyState
 
 abstract class Ontogeny {
 
-  def getState(age: Double): OntogenyState
+  def getState(age: Int): OntogenyState
 
 }
